+++ R/AddInsetMap.R
@@ -4,7 +4,7 @@
 #'
 #' @param p 'SpatialPolygons'.
 #'   Polygon describing the large map.
-#' @param col 'list'.
+#' @param col 'character'.
 #'   Vector of length 2 giving the colors for filling the large map polygon \code{p} and the smaller plot extent rectangle.
 #' @param main.label 'list'.
 #'   List with components \code{label} and \code{adj}.
@@ -79,7 +79,17 @@ AddInsetMap <- function(p, col=c("#D8D8D8", "#BFA76F"),
   }
   dy <- dx * (diff(ext[3:4]) / diff(ext[1:2]))
 
+
+
+
+
+
   xy <- GetInsetLocation(dx, dy, loc=loc, inset=inset)
+
+
+
+
+
   graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, col="#FFFFFFE7", border=NA)
 
   plt <- c(graphics::grconvertX(c(xy[1], xy[1] + dx), "user", "nfc"),
@@ -106,100 +116,4 @@ AddInsetMap <- function(p, col=c("#D8D8D8", "#BFA76F"),
   if (bty != "n") graphics::box(lwd=0.5)
 
   invisible(NULL)
-}
-
-
-#' Get Inset Location
-#'
-#' This function determines the location for an inset in the main plot region.
-#'
-#' @param dx,dy 'numeric'.
-#'   Width and height of the inset graphics, respectively.
-#' @param loc 'character'.
-#'   Single keyword used to specify the position of the inset in the main plot region:
-#'   \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"},
-#'   \code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"},
-#'   or \code{"center"} to denote inset location.
-#' @param inset 'numeric'.
-#'   Inset distance(s) from the margins as a fraction of the main plot region.
-#'   Defaults to 2 percent of the axis range.
-#' @param padx,pady 'numeric'.
-#'   Padding distance in the \emph{x} and \emph{y} direction, respectively.
-#'
-#' @return Returns a 'numeric' vector of length 2 giving the user coordinates
-#'   for the bottom-left corner of the inset.
-#'
-#' @author J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
-#'
-#' @keywords internal
-#'
-#' @export
-#'
-#' @examples
-#' graphics::plot(NA, NA, xlim = c(0, 100), ylim = c(0, 100),
-#'                xlab = "x", ylab = "y", xaxs = "i", yaxs = "i")
-#' dx <- 20; dy <- 10
-#' xy <- GetInsetLocation(dx, dy)
-#' graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "red")
-#' graphics::points(xy[1], xy[2], pch = 16)
-#'
-#' xy <- GetInsetLocation(dx, dy, inset = 0.05)
-#' graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "pink")
-#' graphics::points(xy[1], xy[2], pch = 16)
-#'
-#' xy <- GetInsetLocation(dx, dy, loc = "topright")
-#' graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "blue")
-#'
-#' xy <- GetInsetLocation(dx, dy, loc = "left", padx = 5)
-#' graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "green")
-#'
-#' xy <- GetInsetLocation(dx, dy, loc = "center")
-#' graphics::rect(xy[1], xy[2], xy[1] + dx, xy[2] + dy, border = "brown")
-#'
-
-GetInsetLocation <- function(dx, dy, loc="bottomleft", inset=0.02, padx=0, pady=0) {
-
-  if (grDevices::dev.cur() == 1) stop("no active device")
-  usr <- graphics::par("usr")
-  w <- diff(usr[1:2])
-  h <- diff(usr[3:4])
-
-  choices <- c("bottomright", "bottom", "bottomleft", "left",
-               "topleft", "top", "topright", "right", "center")
-  loc <- match.arg(loc, choices)
-  dx <- checkmate::assertNumber(dx, lower=0, upper=w, finite=TRUE)
-  dy <- checkmate::assertNumber(dy, lower=0, upper=h, finite=TRUE)
-  inset <- checkmate::assertNumeric(inset, lower=0, upper=1, finite=TRUE,
-                                    any.missing=FALSE, min.len=1, max.len=2)
-  padx <- checkmate::assertNumber(padx, finite=TRUE)
-  pady <- checkmate::assertNumber(pady, finite=TRUE)
-
-  inset <- rep(inset, length.out=2)
-  padx <- padx + inset[1] * w
-  pady <- pady + inset[2] * h
-
-  center <- c(usr[1] + w / 2 - dx / 2, usr[3] + h / 2 - dy / 2)
-
-  if (loc == "bottomright") {
-    xy <- c(usr[2] - padx - dx, usr[3] + pady)
-  } else if (loc == "bottom") {
-    xy <- c(center[1], usr[3] + pady)
-  } else if (loc == "bottomleft") {
-    xy <- c(usr[1] + padx, usr[3] + pady)
-  } else if (loc == "left") {
-    xy <- c(usr[1] + padx, center[2])
-  } else if (loc == "topleft") {
-    xy <- c(usr[1] + padx, usr[4] - pady - dy)
-  } else if (loc == "top") {
-    xy <- c(center[1], usr[4] - pady - dy)
-  } else if (loc == "topright") {
-    xy <- c(usr[2] - padx - dx, usr[4] - pady - dy)
-  } else if (loc == "right") {
-    xy <- c(usr[2] - padx - dx, center[2])
-  } else if (loc == "center") {
-    xy <- center
-  }
-  names(xy) <- c("x", "y")
-
-  return(xy)
 }